feat(settings): merge UI, privacy, language and icon into Preferences
Do not add a fourth stats control. Move "by ingestion time" into a
new Preferences screen with the existing UI bools (24h, dosage dots,
open in browser, hide timeline, independent heights, midnight cutoff),
plus app lock, effect notifications, language, and icon picker.

Settings home keeps catalog/data/feedback entries and a Preferences
row. Stats overview still follows the DataStore flag in captions.
