+++ app/src/main/assets/lang/en_us.json
@@ -17,11 +17,15 @@
   "settings_language_title": "Choose language",
   "settings_language_system": "From system",
   "settings_ui": "UI",
+  "settings_preferences": "Preferences",
+  "settings_appearance": "Appearance",
   "settings_custom_units": "Custom units",
   "settings_substance_colors": "Substance colors",
   "settings_interaction_settings": "Interaction settings",
   "settings_hide_dosage_dots": "Hide dosage dots",
   "settings_use_24_hour_clock": "Use 24-hour time",
+  "settings_stats_by_ingestion_time": "Stats by ingestion time",
+  "settings_stats_by_ingestion_time_description": "Place each ingestion on the day it happened instead of grouping a whole experience into its first day",
   "settings_app_data": "App data",
   "settings_export_file": "Export file",
   "settings_export_title": "Export?",
