+++ app/src/main/assets/lang/zh_cn.json
@@ -17,11 +17,15 @@
   "settings_language_title": "选择语言",
   "settings_language_system": "跟随系统",
   "settings_ui": "界面",
+  "settings_preferences": "偏好设置",
+  "settings_appearance": "外观",
   "settings_custom_units": "自定义单位",
   "settings_substance_colors": "物质颜色",
   "settings_interaction_settings": "相互作用设置",
   "settings_hide_dosage_dots": "隐藏剂量圆点",
   "settings_use_24_hour_clock": "使用24小时制",
+  "settings_stats_by_ingestion_time": "按摄入时间统计",
+  "settings_stats_by_ingestion_time_description": "每次摄入按真实发生日期计入，而不是把整段体验算在第一天",
   "settings_app_data": "应用数据",
   "settings_export_file": "导出文件",
   "settings_export_title": "导出？",
