+++ app/src/main/assets/lang/zh_tw.json
@@ -17,11 +17,15 @@
   "settings_language_title": "選擇語言",
   "settings_language_system": "依系統",
   "settings_ui": "介面",
+  "settings_preferences": "偏好設定",
+  "settings_appearance": "外觀",
   "settings_custom_units": "自訂單位",
   "settings_substance_colors": "物質顏色",
   "settings_interaction_settings": "相互作用設定",
   "settings_hide_dosage_dots": "隱藏劑量圓點",
   "settings_use_24_hour_clock": "使用24小時制",
+  "settings_stats_by_ingestion_time": "依攝入時間統計",
+  "settings_stats_by_ingestion_time_description": "每次攝入依真實發生日期計入，而不是把整段體驗算在第一天",
   "settings_app_data": "應用程式資料",
   "settings_export_file": "匯出檔案",
   "settings_export_title": "匯出？",
